WIP - split home/lock-screen into model-specific files to save flash space
